第4章のサンプルコードのアップデート
