+++ scripts/chapter4/4-3-2_prompt_template_api.py
