+++ scripts/chapter4/4-3-2_prompt_template_multiple.py
@@ -1,4 +1,4 @@
-from langchain.prompts import PromptTemplate
+from langchain_core.prompts import PromptTemplate
 
 template = """
 あなたは翻訳アシスタントです。
