+++ scripts/chapter4/4-3-3_pydantic_output_parser.py
@@ -28,7 +28,7 @@ model = ChatOpenAI(
     api_key=os.getenv('OPENAI_API_KEY')
 )
 
-# Chain: プロンプト → モデル → パーサ
+# Chain: プロンプト → モデル → パーサー
 chain = prompt | model | parser
 
 # 実行
