+++ scripts/chapter4/4-3-3_str_output_parser.py
@@ -17,10 +17,10 @@ model = ChatOpenAI(
     api_key=os.getenv('OPENAI_API_KEY')
 )
 
-# 出力パーサ
+# 出力パーサー
 parser = StrOutputParser()
 
-# Chain: プロンプト → モデル → パーサ
+# Chain: プロンプト → モデル → パーサー
 chain = prompt | model | parser
 
 target_text = """
