+++ scripts/chapter4/4-3-5_memory.py
