--- scripts/chapter4/4-5-6_tool.py
@@ -1,29 +1,0 @@
-import os
-from dotenv import load_dotenv
-from langchain_core.tools import tool
-from langchain_openai import ChatOpenAI
-from langchain.agents import create_tool_calling_agent, AgentExecutor
-from langchain_core.prompts import ChatPromptTemplate
-
-load_dotenv()
-
-# ツール定義
-@tool
-def multiply(a: int, b: int) -> int:
-    """二つの整数を掛け算します"""
-    return a * b
-
-tools = [multiply]
-llm = ChatOpenAI(model="gpt-5-nano", api_key=os.getenv("OPENAI_API_KEY"))
-
-prompt = ChatPromptTemplate.from_messages([
-    ("system", "与えられたメッセージに従って計算をしてください"),
-    ("user", "{input}"),
-    ("placeholder", "{agent_scratchpad}"),
-])
-
-agent = create_tool_calling_agent(llm, tools, prompt)
-executor = AgentExecutor(agent=agent, tools=tools, verbose=True)
-
-result = executor.invoke({"input": "7に8を掛けてください"})
-print(result["output"])
